requirements.txt
requests==2.28.2
python-dotenv==1.0.0
pymongo==4.3.3
python-telegram-bot==20.1
Pillow==9.4.0
PySocks==1.7.1